requirements.txt
torch
numpy
requests
streamlit
transformers
Pillow